Migrate old objects values to the OIDC class #41
* name refactoring
* added and modified comments
